First Upload to GitHub for freshman project
The basic algorithm to trying to solve the question: How to select different GPUs such that it can be more cost-efficient than one Tesla v100 GPU that is used for machine learning for web applications. Although this is not a very efficient algorithm is still has starter ideas and methods to base later algorithms on.